requirements.txt
python-dotenv
streamlit 
pymupdf 
requests 
beautifulsoup4 
pydantic 
educhain
openai
